+++ R/internal.R
@@ -96,29 +96,6 @@ align_text <- function(x, n) {
               fixed = TRUE))
 }
 
-#' Default solver name
-#'
-#' This function returns the name of the default solver. If no sovlers are
-#' detected on the system, then a `NULL` object is retured.
-#'
-#' @details This function tests if any of the following packages are installed:
-#'   \pkg{Rsymphony}, \pkg{lpsymphony}, \pkg{gurobi}.
-#'
-#' @return `character` indicating the name of the default solver.
-#'
-#' @noRd
-default_solver_name <- function() {
-  if (requireNamespace("gurobi", quietly = TRUE)) {
-    return("gurobi")
-  } else if (requireNamespace("Rsymphony", quietly = TRUE)) {
-    return("Rsymphony")
-  } else if (requireNamespace("lpsymphony", quietly = TRUE)) {
-    return("lpsymphony")
-  } else {
-    return(NULL)
-  }
-}
-
 #' Any solvers installed?
 #'
 #' Test if any solvers are installed.